refactor(rankings): drop the /rankings hub, 308 to /rankings/players
The hub page duplicated the four dedicated rankings with no content of
its own. /rankings now permanently redirects to the reference ranking,
header/footer link straight to it, and the sitemap only lists the
dedicated pages.
