+++ frontend/messages/es/rankings.json
@@ -1,15 +1,5 @@
 {
   "eyebrow": "Clasificaciones",
-  "title": "Clasificaciones de servidores de Minecraft",
-  "description": "Descubre los mejores servidores de Minecraft: los más jugados ahora mismo, los que más rápido crecen, los récords históricos de afluencia y los últimos servidores añadidos. Clasificaciones actualizadas continuamente con datos reales de jugadores.",
-  "viewAll": "Ver la clasificación completa",
-  "meta": {
-    "title": "Clasificaciones de servidores de Minecraft — Top, tendencias y récords",
-    "description": "Las clasificaciones de servidores de Minecraft: top de jugadores en línea, servidores en tendencia, récords de pico histórico y servidores nuevos. Datos en tiempo real, actualizados cada 10 minutos.",
-    "keywords": "clasificación servidor minecraft, mejor servidor minecraft, top servidor minecraft, servidor minecraft popular, servidor minecraft tendencia",
-    "ogTitle": "Clasificaciones de servidores de Minecraft",
-    "ogDescription": "Top de jugadores en línea, servidores en tendencia, récords de pico y servidores nuevos de Minecraft — en tiempo real."
-  },
   "sections": {
     "players": {
       "nav": "Top jugadores",
